Upgrade to Rust edition 2024 and fix compilation errors
- Update jxl_oxide API: use image_all_channels().buf() instead of buf()
- Fix wgpu-profiler API changes (remove profiling calls)
- Fix usvg Tree::from_data API (2 args instead of 3)
- Fix quick_cache Cache::new type (usize instead of u64)
- Fix jxl_oxide Render API (use FrameBuffer)
- Fix render_loading method signature (&mut self)
- Fix borrow issues in next_image/prev_image
- Fix zoom_by pattern (remove ref keyword)
- Add transition_factor to RenderParams
- Fix needless_option_as_deref clippy warning
- Add #[allow(dead_code)] for profiler and thumb_target_scroll
- Fix unnecessary casts in JXL loader

+++ src/app/state.rs
@@ -21,6 +21,7 @@ pub struct NavigationState {
     pub(crate) load_generation: Arc<AtomicU64>,
     pub(crate) thumb_scroll: f32,
     pub(crate) thumb_velocity: f32,
+    #[allow(dead_code)]
     pub(crate) thumb_target_scroll: f32,
 }
 
@@ -77,7 +78,7 @@ impl SpedImageApp {
             navigation: NavigationState {
                 held_key: None,
                 last_advance_time: None,
-                prefetch_cache: Arc::new(Cache::new(constants::PREFETCH_CACHE_SIZE)),
+                prefetch_cache: Arc::new(Cache::new(constants::PREFETCH_CACHE_SIZE as usize)),
                 load_generation: Arc::new(AtomicU64::new(0)),
                 thumb_scroll: 0.0,
                 thumb_velocity: 0.0,
